Great work. I went to your quickstart with someone new at Teradata and angular/cli and saw some items which need to be changed in order to follow the angular/cli implementation more explicitly. Basically means settin mock api path into environment and moving services to app structure.
