+++ src/app/dashboard-product/overview/overview.component.ts
@@ -3,7 +3,7 @@ import { Title }     from '@angular/platform-browser';
 import { single, multi, multi2 } from './data';
 import { TdLoadingService, TdDigitsPipe } from '@covalent/core';
 
-import { ItemsService, UsersService } from '../../../services';
+import { ItemsService, UsersService } from '../../services';
 
 @Component({
   selector: 'qs-product-overview',
