+++ src/app/services/items.service.ts
@@ -1,14 +1,15 @@
 import { Injectable } from '@angular/core';
 import { Response } from '@angular/http';
 import { HttpInterceptorService, RESTService } from '@covalent/http';
-import { MOCK_API } from '../config/api.config';
+
+import { environment } from '../../environments/environment';
 
 @Injectable()
 export class ItemsService extends RESTService<any> {
 
   constructor(private _http: HttpInterceptorService) {
     super(_http, {
-      baseUrl: MOCK_API,
+      baseUrl: environment.rootApi,
       path: '/items',
     });
   }
