--- src/config/api.config.ts
@@ -1,1 +1,0 @@
-export const MOCK_API: string = 'http://localhost:8080';
