+++ src/environments/environment.prod.ts
@@ -1,3 +1,6 @@
-export const environment: {production: boolean} = {
-  production: true,
+export const environment: {
+  production: boolean,
+  rootApi: string} = {
+    production: true,
+    rootApi: 'http://localhost:8080',
 };
