+++ src/environments/environment.ts
@@ -3,6 +3,9 @@
 // `ng build --env=prod` then `environment.prod.ts` will be used instead.
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
-export const environment: {production: boolean} = {
-  production: false,
+export const environment: {
+  production: boolean,
+  rootApi: string} = {
+    production: false,
+    rootApi: 'http://localhost:8080',
 };
